# api/openapi.yaml
openapi: 3.0.3
info:
  title: ME CLI Sunset API
  description: |
    REST API for managing XL Axiata mobile accounts, data packages, family plans,
    and circle groups. Converted from the original Python CLI application to a
    professional Go API service with SQLite credential storage.

    ## Repository
    - **Source**: [github.com/kertasbaru/me-cli-sunset](https://github.com/kertasbaru/me-cli-sunset)

    ## Authentication Flow
    1. Request OTP via `POST /api/v1/auth/otp/request`
    2. Submit OTP via `POST /api/v1/auth/otp/submit`
    3. Use the active session for subsequent API calls

    ## Storage
    All credentials and session data are stored in a local SQLite database.
  version: 1.0.0
  contact:
    name: ME CLI Sunset
    url: https://github.com/kertasbaru/me-cli-sunset
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/auth/otp/request:
    post:
      tags: [Authentication]
      summary: Request OTP
      description: |
        Sends an OTP code to the specified phone number via SMS.
        The phone number must start with `628` (Indonesian format).
      operationId: requestOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/auth/otp/submit:
    post:
      tags: [Authentication]
      summary: Submit OTP
      description: |
        Verifies the OTP code and authenticates the user.
        On success, the account is automatically saved and set as active.
      operationId: submitOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPSubmitRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
...
  /api/v1/auth/accounts/switch:
    post:
      tags: [Authentication]
      summary: Switch active account
      description: Sets a different stored account as the active one.
      operationId: switchAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [number]
              properties:
                number:
                  type: integer
                  format: int64
                  description: Phone number of the account to activate
                  example: 6281234567890
      responses:
        "200":
          description: Account switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/accounts/{number}:
    delete:
      tags: [Authentication]
      summary: Delete account
      description: Removes a stored account from the database.
      operationId: deleteAccount
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Phone number of the account to delete
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/auth/active:
    get:
      tags: [Authentication]
      summary: Get active user
      description: Returns information about the currently active user.
      operationId: getActiveUser
      responses:
        "200":
          description: Active user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: No active user
  /api/v1/auth/renew:
    post:
      tags: [Authentication]
      summary: Renew token
      description: Forces a token renewal for the active session.
      operationId: renewToken
      responses:
        "200":
          description: Token renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Get notifications
      description: Fetches all notifications for the active user.
      operationId: getNotifications
      responses:
        "200":
          description: Notifications data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/{notification_id}:
    get:
      tags: [Notifications]
      summary: Get notification detail
      description: Fetches detail for a specific notification.
      operationId: getNotificationDetail
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/transactions:
    get:
      tags: [Notifications]
      summary: Get transaction history
      description: Fetches the transaction history for the active user.
      operationId: getTransactionHistory
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/dashboard:
    get:
      tags: [Notifications]
      summary: Get dashboard
      description: Fetches dashboard segment data for the active user.
      operationId: getDashboard
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/bookmarks:
    get:
      tags: [Bookmarks]
      summary: List bookmarks
      description: Returns all bookmarks for the active user from the SQLite database.
      operationId: listBookmarks
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Bookmarks]
      summary: Add bookmark
      description: Adds a new package bookmark for the active user.
      operationId: addBookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkRequest"
      responses:
        "200":
          description: Bookmark added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/bookmarks/{id}:
    delete:
      tags: [Bookmarks]
      summary: Delete bookmark
      description: Removes a bookmark by its ID.
      operationId: deleteBookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Bookmark deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          description: Error message
    OTPRequest:
      type: object
      required: [contact]
      properties:
        contact:
          type: string
          description: Phone number in Indonesian format (starts with 628)
          example: "6281234567890"
    OTPSubmitRequest:
      type: object
      required: [contact, code]
      properties:
        contact_type:
          type: string
          description: Contact type (SMS or DEVICEID)
          default: SMS
        contact:
          type: string
          description: Phone number
          example: "6281234567890"
        code:
          type: string
          description: 6-digit OTP code
          example: "123456"
...
    BookmarkRequest:
      type: object
      required: [family_code]
      properties:
        family_code:
          type: string
          description: Package family code
        family_name:
          type: string
          description: Package family name
        is_enterprise:
          type: boolean
          default: false
        variant_name:
          type: string
          description: Package variant name
        option_name:
          type: string
          description: Package option name
        order:
          type: integer
          default: 0
          description: Display order
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: No active session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"